fix resolver entry
