+++ packages/subgraph/src/mappings/resolver.ts
@@ -1,18 +1,13 @@
-import { Bytes, ethereum } from "@graphprotocol/graph-ts";
+import { Address, Bytes, ethereum } from "@graphprotocol/graph-ts";
+import { RoleAdminChanged, RoleGranted, RoleRevoked, Set } from "../../generated/ResolverV1/Resolver";
 import {
-    RoleAdminChanged,
-    RoleGranted,
-    RoleRevoked,
-    Set,
-} from "../../generated/ResolverV1/Resolver";
-import {
+    ResolverEntry,
     RoleAdminChangedEvent,
     RoleGrantedEvent,
     RoleRevokedEvent,
     SetEvent,
     Token,
 } from "../../generated/schema";
-import { getOrInitResolverEntry } from "../mappingHelpers";
 import { createEventID, initializeEventEntity, ZERO_ADDRESS } from "../utils";
 
 export function handleRoleAdminChanged(event: RoleAdminChanged): void {
@@ -47,14 +42,34 @@ export function handleRoleRevoked(event: RoleRevoked): void {
     ev.save();
 }
 
+function getOrInitResolverEntry(id: string, target: Address, block: ethereum.Block): ResolverEntry {
+    let resolverEntry = ResolverEntry.load(id);
+
+    if (resolverEntry == null) {
+        resolverEntry = new ResolverEntry(id);
+        resolverEntry.createdAtBlockNumber = block.number;
+        resolverEntry.createdAtTimestamp = block.timestamp;
+        resolverEntry.targetAddress = target;
+    }
+    const isListed = target.notEqual(ZERO_ADDRESS);
+
+    // we only update this if the target is not equal to the zero address
+    if (isListed) {
+        resolverEntry.isToken = Token.load(target.toHex()) != null;
+    }
+    resolverEntry.updatedAtBlockNumber = block.number;
+    resolverEntry.updatedAtTimestamp = block.timestamp;
+    resolverEntry.isListed = isListed;
+
+    resolverEntry.save();
+
+    return resolverEntry as ResolverEntry;
+}
+
 export function handleSet(event: Set): void {
     _createSetEvent(event, event.params.target, event.params.name);
 
-    const resolverEntry = getOrInitResolverEntry(
-        event.params.name.toHex(),
-        event.params.target,
-        event.block
-    );
+    const resolverEntry = getOrInitResolverEntry(event.params.name.toHex(), event.params.target, event.block);
 
     // upon initial setting, we will know if this address belongs to a token contract
     if (resolverEntry.isToken) {
@@ -80,11 +95,7 @@ export function handleSet(event: Set): void {
     resolverEntry.save();
 }
 
-function _createSetEvent(
-    event: ethereum.Event,
-    target: Bytes,
-    name: Bytes
-): void {
+function _createSetEvent(event: ethereum.Event, target: Bytes, name: Bytes): void {
     const eventId = createEventID("Set", event);
     const ev = new SetEvent(eventId);
     initializeEventEntity(ev, event, [target]);
